Show achievement icons and make a partial sync resumable
Fetch the game's achievement list from GetSchemaForGame rather than
deriving it from a player's response: it carries the official icons and
the hidden flag, has no user context, and is readable regardless of
profile privacy. Being identical for everyone, it is cached once per game
and costs nothing for every user after.

Steam enforces an undocumented request budget per window, so a full
library can exhaust it part-way through. Games synced in the last few
hours are now skipped without a request, which lets a retry finish the
remainder instead of spending the budget again on work already done.
Throttling is reported as its own outcome, since progress is kept and the
answer is to run again shortly.

Hidden achievements show their description only once earned.

+++ backend/src/main/java/dev/nexus/modules/games/OutboundRateLimiter.java
@@ -13,8 +13,8 @@ public class OutboundRateLimiter {
     private final long spacingNanos;
     private long nextPermitAt;
 
-    public OutboundRateLimiter(int requestsPerSecond) {
-        this.spacingNanos = TimeUnit.SECONDS.toNanos(1) / requestsPerSecond;
+    public OutboundRateLimiter(double requestsPerSecond) {
+        this.spacingNanos = (long) (TimeUnit.SECONDS.toNanos(1) / requestsPerSecond);
         this.nextPermitAt = System.nanoTime();
     }
